docs+chore: week-0 excellence month stop (S+ 100, ROADMAP truth, ruff)
WP-A: early entity blockquote + markdown TOC for S+ overall 100.
WP-B: post-1.13 ROADMAP Current/Next/Human/Archive structure.
WP-G: ruff --fix plus remaining safe cleanups on watching/te.

+++ kinocut/te/seek.py
@@ -28,5 +28,5 @@ def timestamp_to_frame(seconds: float, fps: float) -> dict[str, Any]:
         raise MCPVideoError("seconds must be >= 0", error_type="validation_error", code="bad_seconds")
     if fps <= 0:
         raise MCPVideoError("fps must be > 0", error_type="validation_error", code="bad_fps")
-    frame = int(round(seconds * fps))
+    frame = round(seconds * fps)  # int via round for half-up on py3
     return frame_to_timestamp(frame, fps)
